requirements.txt
yfinance==0.2.54
kafka-python==2.0.2
pyspark==3.5.1
scikit-learn==1.4.2
streamlit==1.35.0
pandas==2.2.2
pyarrow==16.0.0
plotly==5.22.0